% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary}
\alias{summary}
\alias{summary.onephase}
\alias{summary.twophase}
\alias{summary.threephase}
\title{Summarizing Global and Small-Area Estimation Results}
\usage{
\method{summary}{onephase}(object, coefs = FALSE, ...)

\method{summary}{twophase}(object, coefs = FALSE, ...)

\method{summary}{threephase}(object, coefs = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{onephase}, \code{twophase} or \code{threephase},
containing estimation results of the respective estimation method.}

\item{coefs}{of type "\code{\link[base]{logical}}". If set to \code{TRUE}, also
gives the regression coefficients of \code{\link{twophase}} and
\code{\link{threephase}} estimations. Defaults to \code{FALSE}.}

\item{...}{additional arguments, so far ignored.}
}
\description{
Summarizing Global and Small-Area Estimation Results
}
