% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mphasegain.R
\name{mphase.gain}
\alias{mphase.gain}
\title{mphase.gain}
\usage{
mphase.gain(esttable.obj, pref.vartype = "g_variance", exclude.synth = TRUE)
}
\arguments{
\item{esttable.obj}{an object of class \code{esttable} created by the \code{\link{estTable}} function}

\item{pref.vartype}{preferred type of multiphase variance that should be compared to the \code{onephase} variance,
if more then one variance type has been calculated in the multiphase estimation object(s) stored in
\code{esttable}. Valid input values are \code{"g_variance"} (default) and \code{"ext_variance"}.}

\item{exclude.synth}{\code{logical}. If set to \code{TRUE} (default), synthetic estimations are not considered in the validation.}
}
\value{
\code{mphase.gain} returns a \code{data.frame} containing the following components:

 \itemize{
    \item \code{area:} in case of small area estimation: the name of the small area
    \item \code{std_onephase:} standard error of the \code{\link{onephase}} estimation
    \item \code{std_multiphase:} smallest standard error among the (set of) multiphase estimations stored in \code{esttable.obj}
    \item \code{estimator:} multiphase estimator with the smallest standard error
    \item \code{method:} estimation Method of the multiphase estimator with the smallest standard error
    \item \code{gain:} the \emph{gain} is the reduction (if value is positive) or possibly also the increase (if value is negative)
                       in standard error when applying the multiphase as alternative to the onephase estimation
    \item \code{perc.of.onephase:} ratio between the smallest multiphase standard error and the onephase standard error
    \item \code{rel.eff:} the \emph{relative efficiency} defined as the ratio between the onephase standard error and the multiphase standard error
 }
}
\description{
\code{mphase.gain} takes as input an object created by the \code{\link{estTable}} function
and returns a validation of which multiphase method and estimator performed best in comparison
to the onephase estimation (baseline) in terms of estimation precision.
}
\note{
The standard error is defined as the square root of the estimated variance

The \emph{gain} can be interpreted as: "The multiphase estimation procedure leads a xx % reduction in standard error compared to the
onephase procedure".

The \emph{relative efficiency} can be interpreted as: "The terrestrial sample size would have to be 'rel.eff' times larger in order to achieve the
mutiphase estimation precision (in terms of standard error) with the onephase estimation procedure".
}
\examples{

## run onephase estimation:
op.a <- onephase(formula = tvol~1,
                 data = grisons,
                 phase_id = list(phase.col = "phase_id_2p", terrgrid.id = 2),
                 area = list(sa.col = "smallarea", areas = c("A", "B", "C", "D")))

## run small area twophase estimation:
sae.2p.est <- twophase(formula = tvol ~ mean + stddev + max + q75,
                       data = grisons,
                       phase_id = list(phase.col = "phase_id_2p", terrgrid.id = 2),
                       small_area = list(sa.col = "smallarea", areas = c("A", "B","C", "D"),
                                         unbiased = TRUE))

sae.2p.est2 <- twophase(formula = tvol ~ mean + stddev + max + q75, data = grisons,
                        phase_id = list(phase.col = "phase_id_2p", terrgrid.id = 2),
                        small_area = list(sa.col = "smallarea", areas = c("A", "B"),
                                          unbiased = TRUE),
                        psmall = TRUE)


## create estimation table:
sae.table<- estTable(est.list = list(op.a, sae.2p.est, sae.2p.est2),
                     sae = TRUE, vartypes = c("variance", "g_variance",  "ext_variance"))


# validate estimation results:
mphase.gain(sae.table)
}
